services_content-generation_1746497988.908000/requirements.txt
functions-framework>=3.0.0
flask>=2.2
google-cloud-secret-manager>=2.16
google-generativeai>=0.4
pandas>=1.5
protobuf>=4.25
grpcio>=1.59
